iox-#1391 Move 'AccessMode' and 'OpenMode' to 'filesystem.hpp'
