+++ iceoryx_hoofs/CMakeLists.txt
@@ -112,7 +112,6 @@ iox_add_library(
         source/posix_wrapper/signal_handler.cpp
         source/posix_wrapper/system_configuration.cpp
         source/posix_wrapper/thread.cpp
-        source/posix_wrapper/types.cpp
         source/posix_wrapper/unnamed_semaphore.cpp
         source/posix_wrapper/unix_domain_socket.cpp
         source/error_reporting/custom/default/default_error_handler.cpp
