+++ iceoryx_hoofs/source/posix_wrapper/file_lock.cpp
@@ -16,11 +16,11 @@
 
 #include "iceoryx_hoofs/posix_wrapper/file_lock.hpp"
 #include "iceoryx_hoofs/posix_wrapper/posix_call.hpp"
-#include "iceoryx_hoofs/posix_wrapper/types.hpp"
 #include "iceoryx_platform/errno.hpp"
 #include "iceoryx_platform/fcntl.hpp"
 #include "iceoryx_platform/stat.hpp"
 #include "iceoryx_platform/unistd.hpp"
+#include "iox/filesystem.hpp"
 #include "iox/logging.hpp"
 
 #include "iceoryx_platform/platform_correction.hpp"
