+++ iceoryx_hoofs/source/posix_wrapper/shared_memory_object/shared_memory.cpp
@@ -17,12 +17,12 @@
 
 #include "iceoryx_hoofs/internal/posix_wrapper/shared_memory_object/shared_memory.hpp"
 #include "iceoryx_hoofs/posix_wrapper/posix_call.hpp"
-#include "iceoryx_hoofs/posix_wrapper/types.hpp"
 #include "iceoryx_platform/fcntl.hpp"
 #include "iceoryx_platform/mman.hpp"
 #include "iceoryx_platform/stat.hpp"
 #include "iceoryx_platform/types.hpp"
 #include "iceoryx_platform/unistd.hpp"
+#include "iox/filesystem.hpp"
 #include "iox/logging.hpp"
 #include "iox/scope_guard.hpp"
