+++ iceoryx_posh/test/moduletests/test_roudi_posix_shm_memory_provider.cpp
@@ -50,8 +50,8 @@ class PosixShmMemoryProvider_Test : public Test
         return !iox::posix::SharedMemoryObjectBuilder()
                     .name(TEST_SHM_NAME)
                     .memorySizeInBytes(8)
-                    .accessMode(iox::posix::AccessMode::READ_ONLY)
-                    .openMode(iox::posix::OpenMode::OPEN_EXISTING)
+                    .accessMode(iox::AccessMode::READ_ONLY)
+                    .openMode(iox::OpenMode::OPEN_EXISTING)
                     .permissions(iox::perms::owner_all)
                     .create()
                     .has_error();
@@ -64,8 +64,7 @@ class PosixShmMemoryProvider_Test : public Test
 TEST_F(PosixShmMemoryProvider_Test, CreateMemory)
 {
     ::testing::Test::RecordProperty("TEST_ID", "9808f2a5-4cd3-49fe-9a19-6e747183141d");
-    PosixShmMemoryProvider sut(
-        TEST_SHM_NAME, iox::posix::AccessMode::READ_WRITE, iox::posix::OpenMode::PURGE_AND_CREATE);
+    PosixShmMemoryProvider sut(TEST_SHM_NAME, iox::AccessMode::READ_WRITE, iox::OpenMode::PURGE_AND_CREATE);
     ASSERT_FALSE(sut.addMemoryBlock(&memoryBlock1).has_error());
     uint64_t MEMORY_SIZE{16};
     uint64_t MEMORY_ALIGNMENT{8};
@@ -82,8 +81,7 @@ TEST_F(PosixShmMemoryProvider_Test, CreateMemory)
 TEST_F(PosixShmMemoryProvider_Test, DestroyMemory)
 {
     ::testing::Test::RecordProperty("TEST_ID", "f864b99c-373d-4954-ac8b-61acc3c9c555");
-    PosixShmMemoryProvider sut(
-        TEST_SHM_NAME, iox::posix::AccessMode::READ_WRITE, iox::posix::OpenMode::PURGE_AND_CREATE);
+    PosixShmMemoryProvider sut(TEST_SHM_NAME, iox::AccessMode::READ_WRITE, iox::OpenMode::PURGE_AND_CREATE);
     ASSERT_FALSE(sut.addMemoryBlock(&memoryBlock1).has_error());
     uint64_t MEMORY_SIZE{16};
     uint64_t MEMORY_ALIGNMENT{8};
@@ -102,8 +100,7 @@ TEST_F(PosixShmMemoryProvider_Test, DestroyMemory)
 TEST_F(PosixShmMemoryProvider_Test, CreationFailedWithAlignmentExceedingPageSize)
 {
     ::testing::Test::RecordProperty("TEST_ID", "6614de7e-0f4c-48ea-bd3c-dd500fa231f2");
-    PosixShmMemoryProvider sut(
-        TEST_SHM_NAME, iox::posix::AccessMode::READ_WRITE, iox::posix::OpenMode::PURGE_AND_CREATE);
+    PosixShmMemoryProvider sut(TEST_SHM_NAME, iox::AccessMode::READ_WRITE, iox::OpenMode::PURGE_AND_CREATE);
     ASSERT_FALSE(sut.addMemoryBlock(&memoryBlock1).has_error());
     uint64_t MEMORY_SIZE{16};
     uint64_t MEMORY_ALIGNMENT{iox::internal::pageSize() + 8U};
